Split taskBMP280 to different file Updated CMakeLists.txt
