+++ main/src/taskbmp.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void taskBMP280(void *pvParameters);
